refactor: HomePage에서 PetRegisterModal 컴포넌트 분리
