+++ src/components/home/petRegisterModal/index.js
@@ -1,0 +1,2 @@
+export { default as PetRegisterModalContent } from "./PetRegisterModalContent";
+export { default as PetRegisterModal } from "./PetRegisterModal";
